Move navigation scaffold out of nav host
