+++ composeApp/src/commonMain/kotlin/io/github/jakubherr/gitfit/presentation/auth/AuthNavigation.kt
@@ -21,7 +21,6 @@ fun NavGraphBuilder.authGraph(
         )
     }
 
-    // TODO skip to dashboard after email is sent successfully
     composable<VerifyEmailRoute> { VerifyEmailScreenRoot { navController.navigate(DashboardRoute) } }
 
     composable<ResetPasswordRoute> { ResetPasswordScreenRoot() }
